fix(e2e): harden group create modal org selection against open-click race
The organization combobox loads its options asynchronously, so the open click
could land before the panel was interactive (during hydration/re-render),
leaving the dropdown closed and no option ever rendering - a 10s timeout that
passed on retry and flaked in CI on both Desktop and Mobile Chrome.

Move the selection logic into a shared selectFirstOrganization() method on the
CreateGroupModal page object that clicks only while the button reports
aria-expanded !== "true" and retries until expanded, so a lost open-click is
retried without toggling an already-open panel shut. The two group-create spec
files now delegate to this single implementation instead of duplicating it.
